% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.roprobit.R
\name{plot.roprobit}
\alias{plot.roprobit}
\title{show diagnostic plot of an estimation object of class 'roprobit'}
\usage{
\method{plot}{roprobit}(x, which = 1:min(4, length(x$coef)))
}
\description{
show diagnostic plot of an estimation object of class 'roprobit'
}
