\name{roprobit-package}
\alias{roprobit-package}
\docType{package}
\title{
\packageTitle{roprobit}
}
\description{
\packageDescription{roprobit}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{roprobit}
\packageIndices{roprobit}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{roprobit}

Maintainer: \packageMaintainer{roprobit}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
