% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roprobit.ML.R
\name{roprobit.ML}
\alias{roprobit.ML}
\title{rank ordered probit with maximum simulated likelihood}
\usage{
roprobit.ML(formula, group.ID, data, na.last = T, nCores = 1,
  initparm = NULL)
}
\arguments{
\item{formula}{rank ~ explanatory variables. lower rank = better}

\item{group.ID}{variable containing the IDs of the decision making units}

\item{data}{a data.frame}

\item{na.last}{=TRUE if missing ranks are interpreted as being among the least preferred}

\item{nCores}{number of parallel cores to use}
}
\description{
This function estimates a rank ordered probit model
}
\examples{
put example here

the estimation is under the IIA assumption
}
