% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roprobit.R
\name{roprobit}
\alias{roprobit}
\title{A rank ordered probit model}
\usage{
roprobit(formula, group.ID, choice.ID = NULL, data, na.last = T,
  demean = F, niter = 500, thin = 10, burnin = 50, InnerIter = 1,
  method = "Gibbs", initparm = NULL, nCores = 1)
}
\arguments{
\item{formula}{rank ~ explanatory variables. lower rank = better}

\item{group.ID}{variable containing the IDs of the decision making units}

\item{choice.ID}{variable identifying the choices - not needed for estimation, only to return laten valuations when ranks are missing}

\item{data}{a data.frame}

\item{na.last}{=TRUE if missing ranks are interpreted as being among the least preferred}

\item{niter}{number of data augmentation iterations}

\item{thin}{thinning parameter}

\item{burnin}{number of discarded samples (in the thinned series)}

\item{method}{'MSL' for maximum simulated likelihood, 'Gibbs' for data augmentation}

\item{nCores}{number of parallel cores to use (only with method='Gibbs')}
}
\description{
This function estimates a rank ordered probit model
}
\examples{
put example here

}
